dp-二维数组，int、long溢出，unsigned long可行

+++ c_code/动态规划/子序列/8-115-子序列数.cpp
@@ -22,7 +22,7 @@ class Solution
 public:
     int numDistinct(string s, string t)
     {
-        vector<vector<int>> dp(s.size() + 1, vector<int>(t.size() + 1, 0));
+        vector<vector<unsigned long>> dp(s.size() + 1, vector<unsigned long>(t.size() + 1, 0));
         for (int i = 0; i <= s.size(); i++)
         {
             dp[i][0] = 1;
